se#554  Fix unstable tags slug. Add tests
