+++ catalog/models.py
@@ -294,7 +294,12 @@ class Tag(models.Model):
             self.slug = slugify(
                 unidecode(self.name.replace('.', '-').replace('+', '-'))
             )
-        tag_is_doubled = self.__class__.objects.filter(slug=self.slug).exists()
+        tag_is_doubled = (
+            self.__class__.objects
+            .filter(slug=self.slug)
+            .exclude(group=self.group)
+            .exists()
+        )
         if tag_is_doubled:
             self.slug = randomize_slug(self.slug)
         super(Tag, self).save(*args, **kwargs)
